Extract Batch 3 core tool handlers into src/tools/
Extract remember, recall, and update MCP tool handlers from src/index.ts
into dedicated modules under src/tools/.

- src/tools/remember.ts: registerRememberTool
- src/tools/recall.ts: registerRecallTool
- src/tools/update.ts: registerUpdateTool
