+++ src/index.ts
@@ -45,6 +45,9 @@ import { registerGetProjectIdentityTool } from "./tools/get-project-identity.js"
 import { registerSetProjectIdentityTool } from "./tools/set-project-identity.js";
 import { registerListMigrationsTool, registerExecuteMigrationTool } from "./tools/migration.js";
 import { registerSetProjectMemoryPolicyTool, registerGetProjectMemoryPolicyTool } from "./tools/policy.js";
+import { registerRememberTool } from "./tools/remember.js";
+import { registerRecallTool } from "./tools/recall.js";
+import { registerUpdateTool } from "./tools/update.js";
 import { suggestAutoRelationships } from "./auto-relate.js";
 import {
   computeHybridScore,
@@ -604,1050 +607,15 @@ registerListMigrationsTool(server, ctx);
 
 registerExecuteMigrationTool(server, ctx);
 
-// ── remember ──────────────────────────────────────────────────────────────────
-server.registerTool(
-  "remember",
-  {
-    title: "Remember",
-    description:
-      "REQUIRES: Call `recall` or `list` first to check whether this memory already exists.\n\n" +
-      "Create a new memory as a markdown note with embeddings for future recall.\n\n" +
-      "Use this when:\n" +
-      "- A decision, preference, bug fix, or durable context should survive beyond this session\n" +
-      "- No existing note already covers the topic\n\n" +
-      "Do not use this when:\n" +
-      "- A memory may already exist; use `recall` first to check\n" +
-      "- You need to change an existing memory; use `update`\n" +
-      "- Several overlapping notes should be merged; use `consolidate`\n\n" +
-      "Returns:\n" +
-      "- The created memory id, scope, vault label, lifecycle, and persistence status\n\n" +
-      "Side effects: writes a note, writes embeddings, git commits, and may push.\n\n" +
-      "Typical next step:\n" +
-      "- Use `relate` if this new memory connects to something recalled earlier.",
-    annotations: {
-      readOnlyHint: false,
-      destructiveHint: false,
-      idempotentHint: false,
-      openWorldHint: true,
-    },
-    inputSchema: z.object({
-      title: z.string().max(500, "Title must be at most 500 characters").describe("Specific, retrieval-friendly title. Prefer the concrete topic or decision, not a vague label."),
-      content: z.string().max(100000, "Content must be at most 100,000 characters").describe(
-        "Markdown note body. Put the key fact, decision, or outcome in the opening lines, then supporting detail. Embeddings weight early content more heavily. " +
-        "Content must pass markdown lint. Auto-fixable issues are fixed automatically. Common unfixable issues: fenced code blocks need a language tag (e.g. use ```text not bare ```), and broken links are rejected. " +
-        "If lint fails, fix the specific issues listed in the error and retry the same call."
-      ),
-      tags: z.array(z.string()).optional().default([]).describe("Optional tags for later filtering. Use a small number of stable, meaningful tags."),
-      lifecycle: z
-        .enum(NOTE_LIFECYCLES)
-        .optional()
-        .describe(
-          "Memory lifetime. Use `temporary` for short-lived working context such as active investigations or transient status. " +
-          "Use `permanent` for durable knowledge such as decisions, fixes, patterns, and preferences. " +
-          "When omitted, defaults based on role: research/plan/review → temporary, decision/summary/reference → permanent."
-        ),
-      role: z
-        .enum(NOTE_ROLES)
-        .optional()
-        .describe(
-          "Optional prioritization hint for the note. Inferred automatically when omitted. " +
-          "Set explicitly for workflow artifacts like research or review notes."
-        ),
-      summary: z.string().optional().describe(
-        "Git commit summary only. Imperative mood, concise, and focused on why the change matters."
-      ),
-      alwaysLoad: z
-        .boolean()
-        .optional()
-        .describe(
-          "When true, this note loads automatically at session start and receives priority in recall and relationship expansion. " +
-          "Use for session anchors and critical context that should always be available."
-        ),
-      cwd: z
-        .string()
-        .optional()
-        .describe(
-          "Absolute project working directory. Pass this whenever the task is related to a repository so routing, search boosting, policy lookup, and vault selection work correctly."
-        ),
-      scope: z
-        .enum(WRITE_SCOPES)
-        .optional()
-        .describe(
-          "Where to store: 'project' writes to the shared project vault visible to all contributors; " +
-          "'global' writes to the private main vault visible only on this machine. " +
-          "When omitted, uses the project's saved policy or defaults to 'project'."
-        ),
-      allowProtectedBranch: z
-        .boolean()
-        .optional()
-        .describe(
-          "One-time override for protected branch checks. " +
-          "When true, remember can commit on a protected branch without changing project policy."
-        ),
-      checkedForExisting: z
-        .boolean()
-        .optional()
-        .describe(
-          "Optional agent hint indicating that `recall` or `list` was already used to check for an existing memory on this topic."
-        ),
-    }),
-    outputSchema: RememberResultSchema,
-  },
-  async ({ title, content, tags, lifecycle, role, summary, alwaysLoad, cwd, scope, allowProtectedBranch = false }) => {
-    await ensureBranchSynced(cwd);
-
-    const project = await resolveProject(cwd);
-    let cleanedContent: string;
-    try {
-      cleanedContent = await cleanMarkdown(content);
-    } catch (err) {
-      if (err instanceof MarkdownLintError) {
-        const message = `Markdown lint issues prevented this note from being stored. Fix the specific lint errors listed below in your content and retry the remember call — the note was NOT stored.\n\n${err.message}`;
-        return {
-          content: [{ type: "text" as const, text: message }],
-          structuredContent: { action: "lint_error", tool: "remember", issues: err.issues } as LintErrorResult,
-          isError: true,
-        };
-      }
-      throw err;
-    }
-    const policy = project ? await configStore.getProjectPolicy(project.id) : undefined;
-    const policyScope = policy?.defaultScope;
-    const projectVaultExists = cwd ? Boolean(await vaultManager.getProjectVaultIfExists(cwd)) : true;
-    const writeScope = resolveWriteScope(scope, policyScope, Boolean(project), projectVaultExists);
-    if (writeScope === "ask") {
-      const unadopted = !projectVaultExists && !policyScope;
-      return { content: [{ type: "text", text: formatAskForWriteScope(project, unadopted) }], isError: true };
-    }
-
-    const protectedBranchCheck = await shouldBlockProtectedBranchCommit({
-      cwd,
-      writeScope,
-      automaticCommit: true,
-      projectLabel: project ? `${project.name} (${project.id})` : "this context",
-      policy,
-      allowProtectedBranch,
-      toolName: "remember",
-    });
-    if (protectedBranchCheck.blocked) {
-      return { content: [{ type: "text", text: protectedBranchCheck.message ?? "Protected branch policy blocked this commit." }], isError: true };
-    }
-
-    const vault = await resolveWriteVault(cwd, writeScope);
-
-    const id = makeId(title);
-    const now = isoDateString(new Date().toISOString());
-
-    const note: Note = {
-      id, title, content: cleanedContent, tags,
-      lifecycle: lifecycle ?? (role ? ROLE_LIFECYCLE_DEFAULTS[role] : undefined) ?? "permanent",
-      ...(role ? { role } : {}),
-      alwaysLoad: alwaysLoad ?? false,
-      project: project?.id,
-      projectName: project?.name,
-      createdAt: now,
-      updatedAt: now,
-      memoryVersion: 1,
-    };
-
-    if (project) {
-      const accessCandidates = getRecentSessionNoteAccesses(project.id)
-        .map((entry) => {
-          const cachedNote = getSessionCachedNote(project.id, entry.vaultPath, entry.noteId)
-            ?? getRecentSessionAccessNote(project.id, entry.vaultPath, entry.noteId);
-          return cachedNote
-            ? { note: cachedNote, accessedAt: entry.accessedAt, accessKind: entry.accessKind, score: entry.score }
-            : null;
-        })
-        .filter((entry): entry is NonNullable<typeof entry> => entry !== null);
-      const autoRelationships = suggestAutoRelationships(note, accessCandidates);
-      if (autoRelationships.length > 0) {
-        note.relatedTo = autoRelationships;
-      }
-    }
-
-    await vault.storage.writeNote(note);
-
-    let embeddingStatus: { status: "written" | "skipped"; reason?: string } = { status: "written" };
-
-    try {
-      const text = await embedTextForNote(vault.storage, note);
-      const vector = await embed(text);
-      await vault.storage.writeEmbedding({ id, model: embedModel, embedding: vector, updatedAt: now });
-    } catch (err) {
-      embeddingStatus = { status: "skipped", reason: getErrorMessage(err) };
-      console.error(`[embedding] Skipped for '${id}': ${err}`);
-    }
-
-    const projectScope = describeProject(project);
-    const commitSummary = summary ?? extractSummary(cleanedContent);
-    const commitBody = formatCommitBody({
-      summary: commitSummary,
-      noteId: id,
-      noteTitle: title,
-      projectName: project?.name,
-      scope: writeScope,
-      tags: tags,
-    });
-    const commitMessage = `remember: ${title}`;
-    const commitFiles = [vaultManager.noteRelPath(vault, id)];
-    const commitStatus = await vault.git.commitWithStatus(commitMessage, commitFiles, commitBody);
-    const pushStatus = commitStatus.status === "committed"
-      ? await pushAfterMutation(vault)
-      : { status: "skipped" as const, reason: "commit-failed" as const };
-    const retry = buildMutationRetryContract({
-      commit: commitStatus,
-      commitMessage,
-      commitBody,
-      files: commitFiles,
-      cwd,
-      vault,
-      mutationApplied: true,
-    });
-    const persistence = buildPersistenceStatus({
-      storage: vault.storage,
-      id,
-      embedding: embeddingStatus,
-      commit: commitStatus,
-      push: pushStatus,
-      commitMessage,
-      commitBody,
-      retry,
-    });
-
-    const vaultLabel = vault.isProject ? " [project vault]" : " [main vault]";
-    const textContent = `Remembered as \`${id}\` [${projectScope}, stored=${writeScope}]${vaultLabel}\n${formatPersistenceSummary(persistence)}`;
-    
-    const structuredContent: RememberResult = {
-      action: "remembered",
-      id,
-      title,
-      project: project ? { id: project.id, name: project.name } : undefined,
-      scope: writeScope,
-      vault: storageLabel(vault),
-      tags: tags || [],
-      lifecycle: note.lifecycle,
-      timestamp: now,
-      persistence,
-    };
-
-    invalidateActiveProjectCache();
-    return {
-      content: [{ type: "text", text: textContent }],
-      structuredContent,
-    };
-  }
-);
-
 registerSetProjectMemoryPolicyTool(server, ctx);
 
 registerGetProjectMemoryPolicyTool(server, ctx);
 
-// ── recall ────────────────────────────────────────────────────────────────────
-server.registerTool(
-  "recall",
-  {
-    title: "Recall",
-    description:
-      "Semantic search over stored memories using embeddings.\n\n" +
-      "Supports opt-in temporal mode (`mode: \"temporal\"`) to enrich top semantic matches with compact git-backed history.\n\n" +
-      "Supports workflow mode (`mode: \"workflow\"`) to prioritize RPIR-style chain reconstruction while retaining compatibility with legacy relationships.\n\n" +
-      "Use this when:\n" +
-      "- You know the topic but not the exact memory id\n" +
-      "- You are starting a session and want relevant prior context\n" +
-      "- You want to check whether a memory already exists before creating another one\n" +
-      "- You explicitly want to inspect how a note evolved over time\n\n" +
-      "Do not use this when:\n" +
-      "- You already know the exact id; use `get`\n" +
-      "- You just want to browse by tags or scope; use `list`\n\n" +
-      "Returns:\n" +
-      "- Ranked memory matches with scores, vault label, tags, lifecycle, and updated time\n" +
-      "- Bounded 1-hop relationship previews automatically attached to top results\n" +
-      "- In temporal mode, optional compact history entries for top matches\n" +
-      "- Optional retrieval evidence via `evidence: \"compact\"` for why a result ranked\n\n" +
-      "Read-only.\n\n" +
-      "Typical next step:\n" +
-      "- Use `get`, `update`, `relate`, or `consolidate` based on the results.",
-    annotations: {
-      readOnlyHint: true,
-      destructiveHint: false,
-      idempotentHint: true,
-      openWorldHint: true,
-    },
-    inputSchema: z.object({
-      query: z.string().describe("Natural-language search query describing the topic, decision, bug, preference, or context you want to find."),
-      cwd: projectParam,
-      limit: z.number().int().min(1).max(20).optional().default(DEFAULT_RECALL_LIMIT),
-      minSimilarity: z.number().min(0).max(1).optional().default(DEFAULT_MIN_SIMILARITY),
-      mode: z.enum(["default", "temporal", "workflow"]).optional().default("default").describe("Use `temporal` for compact git-backed history, or `workflow` for RPIR-oriented chain reconstruction."),
-      verbose: z.boolean().optional().default(false).describe("Only meaningful with `mode: \"temporal\"`. Adds richer stats-based history context without returning raw diffs."),
-      evidence: z.enum(["compact"]).optional().describe("Optional retrieval rationale. Omit for default output; use `compact` for bounded rank and lineage signals."),
-      tags: z.array(z.string()).optional().describe("Filter results to notes with all of these tags."),
-      scope: z
-        .enum(["project", "global", "all"])
-        .optional()
-        .default("all")
-        .describe(
-          "'project' = only this project's memories (project-scoped storage), " +
-          "'global' = only unscoped memories (main/global storage), " +
-          "'all' = both, with project notes boosted (default)"
-        ),
-      lifecycle: z.enum(["temporary", "permanent"]).optional().describe("Filter results by lifecycle. Useful for recovering working-state with `lifecycle: temporary` after `project_memory_summary` orientation."),
-    }),
-    outputSchema: RecallResultSchema,
-  },
-  async ({ query, cwd, limit, minSimilarity, mode, verbose, evidence, tags, scope, lifecycle }) => {
-    const t0Recall = performance.now();
-    await ensureBranchSynced(cwd);
+registerRememberTool(server, ctx);
 
-    const project = await resolveProject(cwd);
-    const queryVec = await embed(query);
-    const vaults = await vaultManager.searchOrder(cwd);
-    const noteCache = new Map<string, Note>();
+registerRecallTool(server, ctx);
 
-    const noteCacheKey = (vault: Vault, id: string): string => `${vault.storage.vaultPath}::${id}`;
-    const readCachedNote = async (vault: Vault, id: string): Promise<Note | null> => {
-      // Check session cache first (populated when getOrBuildVaultEmbeddings was called)
-      if (project) {
-        const sessionNote = getSessionCachedNote(project.id, vault.storage.vaultPath, id);
-        if (sessionNote !== undefined) return sessionNote;
-      }
-      const key = noteCacheKey(vault, id);
-      const cached = noteCache.get(key);
-      if (cached) {
-        return cached;
-      }
-
-      const note = await vault.storage.readNote(memoryId(id));
-      if (note) {
-        noteCache.set(key, note);
-      }
-
-      return note;
-    };
-
-    await Promise.allSettled(
-      vaults.map((vault) =>
-        embedMissingNotes(vault.storage).catch(() => { /* best-effort: don't block recall if Ollama is down */ }),
-      ),
-    );
-
-    const scored: ScoredRecallCandidate[] = [];
-    const temporalQueryHint = detectTemporalQueryHint(query);
-    const applyTemporalFilter = shouldApplyTemporalFiltering(temporalQueryHint);
-    const temporalFilterWindowDays = applyTemporalFilter ? temporalQueryHint?.filterWindowDays : undefined;
-
-    for (const vault of vaults) {
-      const embeddings = project
-        ? (await getOrBuildVaultEmbeddings(project.id, vault)) ?? await vault.storage.listEmbeddings()
-        : await vault.storage.listEmbeddings();
-
-      for (const rec of embeddings) {
-        const rawScore = cosineSimilarity(queryVec, rec.embedding);
-        if (rawScore < minSimilarity) continue;
-
-        const note = await readCachedNote(vault, rec.id);
-        if (!note) continue;
-
-        if (tags && tags.length > 0) {
-          const noteTags = new Set(note.tags);
-          if (!tags.every((t) => noteTags.has(t))) continue;
-        }
-
-        if (lifecycle && note.lifecycle !== lifecycle) {
-          continue;
-        }
-
-        const isProjectNote = note.project !== undefined;
-        const isCurrentProject = project && note.project === project.id;
-
-        if (scope === "project") {
-          if (!isCurrentProject) continue;
-        } else if (scope === "global") {
-          if (isProjectNote) continue;
-        }
-
-        if (
-          applyTemporalFilter
-          && temporalFilterWindowDays !== undefined
-          && !isWithinTemporalFilterWindow(note.updatedAt, temporalFilterWindowDays)
-        ) {
-          continue;
-        }
-
-        const context = buildRecallCandidateContext(note);
-        const temporalBoost = temporalQueryHint
-          ? computeTemporalRecencyBoost(note.updatedAt, temporalQueryHint)
-          : 0;
-        const boost = (isCurrentProject ? PROJECT_SCOPE_BOOST : 0) + context.metadataBoost + temporalBoost;
-        scored.push({
-          id: rec.id,
-          score: rawScore,
-          semanticScoreForPromotion: rawScore,
-          boosted: rawScore + boost,
-          vault,
-          isCurrentProject: Boolean(isCurrentProject),
-          lifecycle: context.lifecycle,
-          relatedCount: context.relatedCount,
-          connectionDiversity: context.connectionDiversity,
-          structureScore: context.structureScore,
-          metadata: context.metadata,
-        });
-      }
-    }
-
-    const projectionTexts = new Map<string, string>();
-    const noteRelationships = new Map<string, Array<{ id: string; type: RelationshipType }>>();
-    for (const candidate of scored) {
-      const note = await readCachedNote(candidate.vault, candidate.id).catch(() => null);
-      if (!note) {
-        continue;
-      }
-
-      if (note.relatedTo && note.relatedTo.length > 0) {
-        noteRelationships.set(candidate.id, note.relatedTo.map((r) => ({ id: r.id, type: r.type })));
-      }
-
-      const projection = await getOrBuildProjection(candidate.vault.storage, note).catch(() => undefined);
-      if (!projection) {
-        continue;
-      }
-
-      projectionTexts.set(candidate.id, projection.projectionText);
-      if (project) {
-        setSessionCachedProjection(project.id, candidate.id, projection);
-      }
-    }
-
-    // Apply lexical reranking over semantic candidates (fail-soft)
-    const getProjectionText = (id: string): string | undefined => {
-      const inlineProjection = projectionTexts.get(id);
-      if (inlineProjection) {
-        return inlineProjection;
-      }
-      if (project) {
-        const cached = getSessionCachedProjection(project.id, id);
-        if (cached) return cached.projectionText;
-      }
-      return undefined;
-    };
-    const strongestSemanticScore = scored.reduce<number | undefined>(
-      (max, candidate) => max === undefined ? candidate.score : Math.max(max, candidate.score),
-      undefined
-    );
-    const reranked = applyLexicalReranking(scored, query, getProjectionText);
-    const semanticCandidateIds = new Set(reranked.map((candidate) => candidate.id));
-
-    // Apply graph spreading activation: traverse related notes and boost their scores
-    const preSpreadIds = new Set(reranked.map((c) => c.id));
-    const getNoteRelationships = (id: string): Array<{ id: string; type: RelationshipType }> | undefined => {
-      return noteRelationships.get(id);
-    };
-    const withGraphSpread = applyGraphSpreadingActivation(reranked, getNoteRelationships);
-    const graphDiscoveredIds = new Set(withGraphSpread.filter((candidate) => !semanticCandidateIds.has(candidate.id)).map((candidate) => candidate.id));
-
-    // Resolve correct vault for graph-discovered candidates that inherited their
-    // entry point's vault instead of their own.
-    await resolveDiscoveredVaults(withGraphSpread, preSpreadIds, async (id) => {
-      for (const v of vaults) {
-        const note = await v.storage.readNote(memoryId(id)).catch(() => null);
-        if (note) {
-          const isCurrentProject = project ? note.project === project.id : false;
-          return { vault: v, isCurrentProject };
-        }
-      }
-      return undefined;
-    });
-
-    // Re-assign semanticRank after graph spreading since scores are now modified
-    // and graph-discovered candidates have no semanticRank.
-    const sortedByScore = [...withGraphSpread].sort((a, b) => b.score - a.score || b.boosted - a.boosted);
-    assignDenseRanks(sortedByScore, (candidate) => candidate.score, (candidate, rank) => {
-      candidate.semanticRank = rank;
-    });
-
-    let promoted = applyCanonicalExplanationPromotion(withGraphSpread);
-    let rescueCandidateIds = new Set<string>();
-
-    // Lexical rescue: when semantic results are weak, scan projections for additional candidates.
-    // Skip rescue when the caller set a strict minSimilarity above the default,
-    // because rescue candidates lack genuine semantic backing.
-    const rescueAllowed = minSimilarity <= DEFAULT_MIN_SIMILARITY;
-    if (rescueAllowed && shouldTriggerLexicalRescue(strongestSemanticScore, scored.length)) {
-      const rescueCandidates = await collectLexicalRescueCandidates(
-        vaults,
-        query,
-        temporalQueryHint,
-        project ?? undefined,
-        scope,
-        tags,
-        lifecycle,
-        promoted
-      );
-      promoted.push(...rescueCandidates);
-      rescueCandidateIds = new Set(rescueCandidates.map((candidate) => candidate.id));
-      enrichRescueCandidateScores(promoted, query, getProjectionText);
-      promoted = applyCanonicalExplanationPromotion(promoted);
-    }
-
-    const top = mode === "workflow"
-      ? selectWorkflowResults(promoted, limit, scope)
-      : selectRecallResults(promoted, limit, scope);
-
-    if (top.length === 0) {
-      const structuredContent: RecallResult = { action: "recalled", query, scope: scope || "all", results: [] };
-      return { content: [{ type: "text", text: "No memories found matching that query." }], structuredContent };
-    }
-
-    const header = project
-      ? `Recall results for project **${project.name}** (scope: ${scope}):`
-      : `Recall results (global):`;
-
-    const sections: string[] = [];
-    const structuredResults: Array<{
-      id: string;
-      title: string;
-      score: number;
-      boosted: number;
-      project?: ProjectRef;
-      vault: string;
-      tags: string[];
-      lifecycle: NoteLifecycle;
-      role?: Note["role"];
-      updatedAt: string;
-      provenance?: {
-        lastUpdatedAt: string;
-        lastCommitHash: string;
-        lastCommitMessage: string;
-        recentlyChanged: boolean;
-      };
-      confidence?: Confidence;
-      history?: Array<{
-        commitHash: string;
-        timestamp: string;
-        message: string;
-        summary?: string;
-        stats?: {
-          additions: number;
-          deletions: number;
-          filesChanged: number;
-          changeType: "metadata-only change" | "minor edit" | "substantial update";
-        };
-      }>; 
-      historySummary?: string;
-      relationships?: RelationshipPreview;
-      retrievalEvidence?: RetrievalEvidence;
-    }> = [];
-    
-    // Determine how many top results get relationship expansion
-    // Top 1 by default, top 3 if result count is small
-    const recallRelationshipLimit = top.length <= 3 ? 3 : 1;
-    
-    for (const [index, { id, score, vault, boosted, semanticRank, lexicalRank, canonicalExplanationScore, metadata, isCurrentProject }] of top.entries()) {
-      const note = await readCachedNote(vault, id);
-      if (note) {
-        const centrality = note.relatedTo?.length ?? 0;
-        const filePath = `${vault.notesRelDir}/${id}.md`;
-        const provenance = await getNoteProvenance(vault.git, filePath);
-        const confidence = computeConfidence(note.lifecycle, note.updatedAt, centrality);
-        let history: Array<{
-          commitHash: string;
-          timestamp: string;
-          message: string;
-          summary?: string;
-          stats?: {
-            additions: number;
-            deletions: number;
-            filesChanged: number;
-            changeType: "metadata-only change" | "minor edit" | "substantial update";
-          };
-        }> | undefined;
-
-        let historySummary: string | undefined;
-
-        if (mode === "temporal") {
-          if (index < TEMPORAL_HISTORY_NOTE_LIMIT) {
-            const commits = await vault.git.getFileHistory(filePath, TEMPORAL_HISTORY_COMMIT_LIMIT);
-            const rawHistory = await Promise.all(
-              commits.map(async (commit) => {
-                const stats = await vault.git.getCommitStats(filePath, commit.hash);
-                return buildTemporalHistoryEntry(commit, stats, verbose);
-              })
-            );
-            const enriched = enrichTemporalHistory(rawHistory);
-            history = verbose
-              ? enriched.interpretedHistory
-              : enriched.interpretedHistory.map(entry => ({ ...entry, stats: undefined }));
-            historySummary = enriched.historySummary;
-          }
-        }
-
-        // Add relationship preview for top N results (fail-soft)
-        let relationships: RelationshipPreview | undefined;
-        if (index < recallRelationshipLimit) {
-          relationships = await getRelationshipPreview(
-            note,
-            vaultManager.allKnownVaults(),
-            { activeProjectId: project?.id, limit: 3 }
-          );
-        }
-
-        const formattedHistory = mode === "temporal" && history !== undefined
-          ? `\n\n${formatTemporalHistory(history)}`
-          : "";
-        const formattedRelationships = relationships !== undefined
-          ? `\n\n${formatRelationshipPreview(relationships)}`
-          : "";
-        const provenanceLine = provenance || confidence
-          ? `\n**confidence:** ${confidence ?? "medium"}${provenance?.recentlyChanged ? " | **recently changed**" : ""}`
-          : "";
-        const supersededRelations = (note.relatedTo ?? []).filter((rel) => rel.type === "supersedes");
-        const retrievalEvidence: RetrievalEvidence | undefined = evidence === "compact"
-          ? {
-            channels: [
-              semanticRank !== undefined ? "semantic" : undefined,
-              lexicalRank !== undefined ? "lexical" : undefined,
-              graphDiscoveredIds.has(id) ? "graph" : undefined,
-              rescueCandidateIds.has(id) ? "rescue" : undefined,
-              canonicalExplanationScore !== undefined && canonicalExplanationScore > 0 ? "canonical" : undefined,
-              temporalQueryHint ? "temporal-boost" : undefined,
-            ].filter((value): value is RetrievalEvidence["channels"][number] => value !== undefined),
-            rankBand: toRecallRankBand(semanticRank),
-            projectRelevant: isCurrentProject,
-            freshness: toRecallFreshness(note.updatedAt),
-            superseded: supersededRelations.length > 0,
-            supersededBy: supersededRelations.length > 0 ? supersededRelations[0]?.id : undefined,
-            supersededCount: supersededRelations.length > 0 ? supersededRelations.length : undefined,
-          }
-          : undefined;
-        const evidenceLine = retrievalEvidence
-          ? `\n${formatRetrievalEvidenceHint(retrievalEvidence, metadata?.role)}`
-          : "";
-        // Suppress raw related IDs when enriched preview is shown to avoid duplication
-        sections.push(`${formatNote(note, score, relationships === undefined)}${provenanceLine}${evidenceLine}${formattedHistory}${formattedRelationships}`);
-
-        structuredResults.push({
-          id,
-          title: note.title,
-          score,
-          boosted,
-          project: noteProjectRef(note),
-          vault: storageLabel(vault),
-          tags: note.tags,
-          lifecycle: note.lifecycle,
-          role: note.role,
-          updatedAt: note.updatedAt,
-          provenance,
-          confidence,
-          history,
-          historySummary,
-          relationships,
-          retrievalEvidence,
-        });
-
-        if (project) {
-          setSessionCachedNote(project.id, vault.storage.vaultPath, note);
-          recordSessionNoteAccess(project.id, vault.storage.vaultPath, id, "recall", computeHybridScore({
-            id,
-            score,
-            boosted,
-            vault,
-            isCurrentProject: note.project === project.id,
-          }));
-        }
-      }
-    }
-
-    const textContent = `${header}\n\n${sections.join("\n\n---\n\n")}`;
-
-    const structuredContent: RecallResult = {
-      action: "recalled",
-      query,
-      scope: scope || "all",
-      results: structuredResults,
-    };
-
-    console.error(`[recall:timing] ${(performance.now() - t0Recall).toFixed(1)}ms`);
-    return {
-      content: [{ type: "text", text: textContent }],
-      structuredContent,
-    };
-  }
-);
-
-// ── update ────────────────────────────────────────────────────────────────────
-server.registerTool(
-  "update",
-  {
-    title: "Update Memory",
-    description:
-      "Use after `recall` + `get` when an existing memory should be refined instead of creating a duplicate.\n\n" +
-      "Modify an existing memory by id.\n\n" +
-      "Use this when:\n" +
-      "- A stored memory is stale, incomplete, or wrong\n" +
-      "- You recalled something useful and want to refine the same note instead of creating a duplicate\n\n" +
-      "Do not use this when:\n" +
-      "- No note exists yet; use `remember`\n" +
-      "- Several notes need to be merged or retired together; use `consolidate`\n\n" +
-      "Returns:\n" +
-      "- The updated memory id, changed fields, and persistence status\n\n" +
-      "Side effects: rewrites the note, refreshes embeddings, git commits, and may push.\n\n" +
-      "Typical next step:\n" +
-      "- Use `relate` or `consolidate` if the update changes how this note connects to others.\n\n" +
-      "Use `semanticPatch` for targeted edits (more token-efficient). Use `content` only for complete rewrites.",
-    annotations: {
-      readOnlyHint: false,
-      destructiveHint: false,
-      idempotentHint: false,
-      openWorldHint: true,
-    },
-    inputSchema: z.object({
-      id: NoteIdSchema.describe("Exact memory id. Use an id returned by `recall`, `list`, `recent_memories`, or `where_is`."),
-      semanticPatch: z
-        .preprocess(
-          (val) => {
-            if (typeof val === "string") {
-              try { return JSON.parse(val); } catch { return val; }
-            }
-            return val;
-          },
-          z.array(z.object({
-          selector: z.object({
-            heading: z.string().optional(),
-            headingStartsWith: z.string().optional(),
-            nthChild: z.number().int().optional(),
-            lastChild: z.literal(true).optional(),
-          }).refine(
-            (sel) => {
-              const keys = [sel.heading, sel.headingStartsWith, sel.nthChild, sel.lastChild].filter((v) => v !== undefined);
-              return keys.length === 1;
-            },
-            { message: "Selector must have exactly one of: heading, headingStartsWith, nthChild, lastChild" }
-          ),
-          operation: z.discriminatedUnion("op", [
-            z.object({ op: z.literal("appendChild"), value: z.string() }),
-            z.object({ op: z.literal("prependChild"), value: z.string() }),
-            z.object({ op: z.literal("replace"), value: z.string() }),
-            z.object({ op: z.literal("replaceChildren"), value: z.string() }),
-            z.object({ op: z.literal("insertAfter"), value: z.string() }),
-            z.object({ op: z.literal("insertBefore"), value: z.string() }),
-            z.object({ op: z.literal("remove") }),
-          ]),
-        })))
-        .optional()
-        .describe(
-          "Targeted edits to note sections. Array of {selector, operation} objects. Mutually exclusive with content. " +
-          "If this fails, fix the issue in your patch values and retry — do NOT fall back to full content rewrite.\n\n" +
-          "selector: exactly one of { heading: \"exact heading text\" } | { headingStartsWith: \"prefix\" } | { lastChild: true } | { nthChild: 0-based-index }\n" +
-          "operation: { op: \"appendChild\", value: \"content\" } | { op: \"prependChild\", value: \"content\" } | { op: \"replace\", value: \"new content\" } | { op: \"replaceChildren\", value: \"new children\" } | { op: \"insertAfter\", value: \"content\" } | { op: \"insertBefore\", value: \"content\" } | { op: \"remove\" }\n\n" +
-          "Example — add a paragraph under ## Findings, replace body under ## Recommendation, remove ## Old Section:\n" +
-          "[\n" +
-          "  { \"selector\": { \"heading\": \"Findings\" }, \"operation\": { \"op\": \"insertAfter\", \"value\": \"A new paragraph.\" } },\n" +
-          "  { \"selector\": { \"heading\": \"Recommendation\" }, \"operation\": { \"op\": \"replace\", \"value\": \"## Recommendation\\n\\nUpdated recommendation.\" } },\n" +
-          "  { \"selector\": { \"heading\": \"Old Section\" }, \"operation\": { \"op\": \"remove\" } }\n" +
-          "]\n\n" +
-          "IMPORTANT: `appendChild`, `prependChild`, and `replaceChildren` do NOT work with `heading` selectors (headings only contain inline text, not block content). To add or replace content under a heading, use `insertAfter`. To replace a heading entirely, use `replace`."
-        ),
-      content: z.string().max(100000, "Content must be at most 100,000 characters").optional().describe("Full note body replacement. Use only for complete rewrites or when the note is small. Mutually exclusive with semanticPatch. Content must pass markdown lint. Auto-fixable issues are fixed automatically. Common unfixable issues: fenced code blocks need a language tag (e.g. use ```text not bare ```), and broken links are rejected. If lint fails, fix the specific issues and retry — do NOT fall back to semanticPatch for this."),
-      title: z.string().max(500, "Title must be at most 500 characters").optional().describe("Specific, retrieval-friendly title. Prefer the concrete topic or decision, not a vague label."),
-      tags: z.array(z.string()).optional().describe("Optional tags for later filtering. Use a small number of stable, meaningful tags."),
-      lifecycle: z
-        .enum(NOTE_LIFECYCLES)
-        .optional()
-        .describe("Change lifecycle. Preserve the existing value unless you're intentionally switching it."),
-      role: z
-        .enum(NOTE_ROLES)
-        .optional()
-        .describe("Change role. Preserve the existing value unless you're intentionally switching it."),
-      summary: z.string().optional().describe("Git commit summary only. Imperative mood, concise, and focused on why the change matters."),
-      alwaysLoad: z
-        .boolean()
-        .optional()
-        .describe(
-          "When true, this note loads automatically at session start and receives priority in recall and relationship expansion. " +
-          "Use for session anchors and critical context that should always be available."
-        ),
-      cwd: projectParam,
-      allowProtectedBranch: z
-        .boolean()
-        .optional()
-        .describe(
-          "One-time override for protected branch checks. " +
-          "When true, update can commit on a protected branch without changing project policy."
-        ),
-    }),
-    outputSchema: UpdateResultSchema,
-  },
-  async ({ id, content, semanticPatch, title, tags, lifecycle, role, summary, alwaysLoad, cwd, allowProtectedBranch = false }) => {
-    await ensureBranchSynced(cwd);
-    const noteId = memoryId(id);
-
-    const found = await vaultManager.findNote(id, cwd);
-    if (!found) {
-      return { content: [{ type: "text", text: `No memory found with id '${id}'` }], isError: true };
-    }
-
-    // Validate: content and semanticPatch are mutually exclusive
-    const hasContent = content !== undefined;
-    const hasSemanticPatch = semanticPatch !== undefined && semanticPatch.length > 0;
-    if (hasContent && hasSemanticPatch) {
-      return { content: [{ type: "text", text: "Exactly one of content or semanticPatch must be provided, not both." }], isError: true };
-    }
-
-    const { note, vault } = found;
-    if (vault.isProject) {
-      const resolvedProject = await resolveProject(cwd);
-      const projectLabel = resolvedProject
-        ? `${resolvedProject.name} (${resolvedProject.id})`
-        : `${note.projectName ?? "project"} (${note.project ?? "unknown"})`;
-      const policy = note.project ? await configStore.getProjectPolicy(note.project) : undefined;
-      const protectedBranchCheck = await shouldBlockProtectedBranchCommit({
-        cwd,
-        writeScope: "project",
-        automaticCommit: true,
-        projectLabel,
-        policy,
-        allowProtectedBranch,
-        toolName: "update",
-      });
-      if (protectedBranchCheck.blocked) {
-        return {
-          content: [{ type: "text", text: protectedBranchCheck.message ?? "Protected branch policy blocked this commit." }],
-          isError: true,
-        };
-      }
-    }
-
-    const now = isoDateString(new Date().toISOString());
-
-    let patchedContent: string | undefined;
-    let lintWarnings: string[] | undefined;
-    if (semanticPatch && semanticPatch.length > 0) {
-      try {
-        const result = await applySemanticPatches(note.content, semanticPatch as SemanticPatch[]);
-        patchedContent = result.content;
-        lintWarnings = result.lintWarnings;
-      } catch (err) {
-        if (err instanceof MarkdownLintError) {
-          const message = `Semantic patch produced content with markdown lint issues. Fix the lint issues in your patch values and retry — do NOT fall back to full content rewrite.\n\n${err.message}`;
-          return { content: [{ type: "text", text: message }], isError: true };
-        }
-        const message = getErrorMessage(err);
-        return { content: [{ type: "text", text: `Semantic patch failed: ${message}` }], isError: true };
-      }
-    }
-    let cleanedContent: string | undefined;
-    if (content !== undefined) {
-      try {
-        cleanedContent = await cleanMarkdown(content);
-      } catch (err) {
-        if (err instanceof MarkdownLintError) {
-          const message = `Markdown lint issues prevented the update. Fix the specific lint errors in your content and retry — do NOT fall back to semanticPatch for this.\n\n${err.message}`;
-          return {
-            content: [{ type: "text" as const, text: message }],
-            structuredContent: { action: "lint_error", tool: "update", issues: err.issues } as LintErrorResult,
-            isError: true,
-          };
-        }
-        throw err;
-      }
-    }
-
-    const resolvedTitle = title ?? note.title;
-    const resolvedContent = patchedContent ?? cleanedContent ?? note.content;
-    const resolvedTags = tags ?? note.tags;
-    const resolvedLifecycle = lifecycle ?? note.lifecycle;
-    const resolvedRole = role !== undefined ? role : (note.role ? note.role : undefined);
-    const resolvedAlwaysLoad = alwaysLoad !== undefined ? alwaysLoad : note.alwaysLoad;
-
-    let relatedToChanged = false;
-    let resolvedRelatedTo = note.relatedTo;
-    if (note.project) {
-      const accessCandidates = getRecentSessionNoteAccesses(note.project)
-        .map((entry) => {
-          const cachedNote = getSessionCachedNote(note.project!, entry.vaultPath, entry.noteId)
-            ?? getRecentSessionAccessNote(note.project!, entry.vaultPath, entry.noteId);
-          return cachedNote
-            ? { note: cachedNote, accessedAt: entry.accessedAt, accessKind: entry.accessKind, score: entry.score }
-            : null;
-        })
-        .filter((entry): entry is NonNullable<typeof entry> => entry !== null);
-      const autoRelationships = suggestAutoRelationships({
-        ...note,
-        title: resolvedTitle,
-        content: resolvedContent,
-        tags: resolvedTags,
-        lifecycle: resolvedLifecycle,
-        alwaysLoad: resolvedAlwaysLoad,
-      }, accessCandidates);
-      if (autoRelationships.length > 0) {
-        const existing = [...(note.relatedTo ?? [])];
-        for (const relationship of autoRelationships) {
-          if (!existing.some((rel) => rel.id === relationship.id && rel.type === relationship.type)) {
-            existing.push(relationship);
-          }
-        }
-        resolvedRelatedTo = existing;
-        relatedToChanged = true;
-      }
-    }
-
-    const changes = computeFieldsModified({
-      patchedContent,
-      originalContent: note.content,
-      contentExplicitlyProvided: content !== undefined,
-      semanticPatchProvided: semanticPatch !== undefined && semanticPatch.length > 0,
-      newTitle: resolvedTitle,
-      originalTitle: note.title,
-      titleExplicitlyProvided: title !== undefined,
-      newLifecycle: resolvedLifecycle,
-      originalLifecycle: note.lifecycle,
-      lifecycleExplicitlyProvided: lifecycle !== undefined,
-      newRole: resolvedRole,
-      originalRole: note.role,
-      roleExplicitlySet: role !== undefined,
-      newTags: resolvedTags,
-      originalTags: note.tags,
-      tagsExplicitlyProvided: tags !== undefined,
-      newAlwaysLoad: resolvedAlwaysLoad,
-      originalAlwaysLoad: note.alwaysLoad,
-      alwaysLoadExplicitlyProvided: alwaysLoad !== undefined,
-      relatedToChanged,
-    });
-
-    const hasChanges = hasActualChanges({
-      content: cleanedContent,
-      originalContent: note.content,
-      title,
-      originalTitle: note.title,
-      tags,
-      originalTags: note.tags,
-      lifecycle,
-      originalLifecycle: note.lifecycle,
-      role,
-      originalRole: note.role,
-      roleExplicitlySet: role !== undefined,
-      alwaysLoad,
-      originalAlwaysLoad: note.alwaysLoad,
-      semanticPatchApplied: semanticPatch !== undefined && semanticPatch.length > 0,
-      relatedToChanged,
-    });
-
-    if (!hasChanges) {
-      const noOpPersistence: PersistenceStatus = {
-        notePath: vault.storage.notePath(memoryId(id)),
-        embeddingPath: vault.storage.embeddingPath(memoryId(id)),
-        embedding: { status: "skipped", model: embedModel, reason: "no-changes" },
-        git: {
-          commit: "skipped",
-          push: "skipped",
-          commitReason: "no-changes",
-          pushReason: "no-changes",
-        },
-        durability: "local-only",
-      };
-      return {
-        content: [{ type: "text", text: `No changes to memory '${id}'` }],
-        structuredContent: {
-          action: "updated" as const,
-          id,
-          title: note.title,
-          fieldsModified: [],
-          timestamp: note.updatedAt,
-          project: noteProjectRef(note),
-          lifecycle: note.lifecycle,
-          role: note.role,
-          persistence: noOpPersistence,
-        },
-      };
-    }
-
-    const updated: Note = {
-      ...note,
-      title: resolvedTitle,
-      content: resolvedContent,
-      tags: resolvedTags,
-      lifecycle: resolvedLifecycle,
-      ...(role !== undefined ? { role: resolvedRole } : (note.role ? { role: note.role } : {})),
-      alwaysLoad: resolvedAlwaysLoad,
-      updatedAt: now,
-      relatedTo: resolvedRelatedTo,
-    };
-
-    await vault.storage.writeNote(updated);
-
-    const shouldReembed = patchedContent !== undefined || cleanedContent !== undefined;
-    let embeddingStatus: { status: "written" | "skipped"; reason?: string } = { status: "skipped", reason: shouldReembed ? undefined : "metadata-only" };
-
-    if (shouldReembed) {
-      try {
-        const text = await embedTextForNote(vault.storage, updated);
-        const vector = await embed(text);
-        await vault.storage.writeEmbedding({ id: noteId, model: embedModel, embedding: vector, updatedAt: now });
-        embeddingStatus = { status: "written" };
-      } catch (err) {
-        embeddingStatus = { status: "skipped", reason: getErrorMessage(err) };
-        console.error(`[embedding] Re-embed failed for '${id}': ${err}`);
-      }
-    }
-
-    const changeDesc = changes.length > 0 ? `Updated ${changes.join(", ")}` : "No changes";
-    const commitSummary = summary ?? changeDesc;
-
-    const commitBody = formatCommitBody({
-      summary: commitSummary,
-      noteId: id,
-      noteTitle: updated.title,
-      projectName: updated.projectName,
-      tags: updated.tags,
-    });
-    const commitMessage = `update: ${updated.title}`;
-    const commitFiles = [vaultManager.noteRelPath(vault, id)];
-    const commitStatus = await vault.git.commitWithStatus(commitMessage, commitFiles, commitBody);
-    const pushStatus = commitStatus.status === "committed"
-      ? await pushAfterMutation(vault)
-      : { status: "skipped" as const, reason: "commit-failed" as const };
-    const retry = buildMutationRetryContract({
-      commit: commitStatus,
-      commitMessage,
-      commitBody,
-      files: commitFiles,
-      cwd,
-      vault,
-      mutationApplied: true,
-    });
-    const persistence = buildPersistenceStatus({
-      storage: vault.storage,
-      id,
-      embedding: embeddingStatus,
-      commit: commitStatus,
-      push: pushStatus,
-      commitMessage,
-      commitBody,
-      retry,
-    });
-
-    const structuredContent: UpdateResult = {
-      action: "updated",
-      id,
-      title: updated.title,
-      fieldsModified: changes,
-      timestamp: now,
-      project: noteProjectRef(updated),
-      lifecycle: updated.lifecycle,
-      role: updated.role,
-      lintWarnings: lintWarnings && lintWarnings.length > 0 ? lintWarnings : undefined,
-      persistence,
-    };
-    
-    invalidateActiveProjectCache();
-    const fieldText = changes.length > 0 ? `\nfields modified: ${changes.join(", ")}` : "";
-    const warningsText = lintWarnings && lintWarnings.length > 0
-      ? `\nmarkdown lint warnings (not auto-fixable):\n- ${lintWarnings.join("\n- ")}`
-      : "";
-    return { content: [{ type: "text", text: `Updated memory '${id}'${fieldText}${warningsText}\n${formatPersistenceSummary(persistence)}` }], structuredContent };
-  }
-);
+registerUpdateTool(server, ctx);
 
 // ── forget ────────────────────────────────────────────────────────────────────
 server.registerTool(
